test(backend): align tests with the fractional-units schema and cash floor
- Switch the per-test `buy` / `transactionCreate` mutation helpers to
  `\$units: Float!` to match the schema change in 8e2d09d
  (`InvestmentTransactions.units` is now `double precision`). The
  schema-side update landed without a sweep over test docs, so every
  test that touched `units` had been failing on this branch.
- Realign the "fully-sold investment" portfolio assertion with the
  cash-floor rule introduced in c2b1776: a wrapper without any
  contribution log doesn't surface realised proceeds as cash, so
  `totalValue` is `0` here (matches the equivalent
  `portfolio-cash.test.ts` assertion); the realised £20 still shows
  via the negative `totalCost`.

+++ packages/backend/src/graphql/investments/investments-page.test.ts
@@ -60,7 +60,7 @@ async function tx(
       $investmentId: ID!
       $assetId: ID!
       $date: Date!
-      $units: Int!
+      $units: Float!
       $priceAmount: Float!
     ) {
       investmentTransactionCreate(
